split mcp_tool and mcp_toolset tests
